Added fixes for GCE issues
GCE was not auto detected. The man pages of kubernetes are unclear on the actual behavior of the auto detect leaning from deprecated to to be implemented.
All the components have been altered to use the gce cloud provider, and have been tested.

TODO:
Load balancer and ingress updates are not yet processed: The glbc needs certificates setup to communicate with the API server.